requirements.txt
fastapi
uvicorn
fastui
python-multipart
mistralai
python-decouple